Split models up into new directory, moved over to migration, split up logic into new services directory, reorganised structure
