+++ app/__init__.py
@@ -1,26 +1,21 @@
+from typing import Type
+
 from flask import Flask
-from flask_wtf.csrf import generate_csrf, CSRFProtect
-from flask_sqlalchemy import SQLAlchemy
-import os
-from app.model import db
-
-basedir = os.path.abspath(os.path.dirname(__file__))
+from config import Config
+from .extensions import db, migrate
 
 
-# Tell Flask where to find the static folder
-app = Flask(
-    __name__,
-    static_folder='../static',
-    template_folder='templates'
-)
+def create_app(configuration: Type[Config]) -> Flask:
+    app = Flask(__name__)
+    app.config.from_object(configuration)
 
-app.secret_key = 'very-secret'
+    db.init_app(app)
+    migrate.init_app(app, db)
 
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + os.path.join(basedir, 'users.db')
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    with app.app_context():
+        from app.models import User, File, Chart, SharedFile, SharedChart
 
-db.init_app(app) 
-csrf = CSRFProtect(app)
+        from app.routes import bp
+        app.register_blueprint(bp)
 
-from app import routes, model
-
+    return app
