+++ app/models/__init__.py
@@ -1,0 +1,4 @@
+from .user import User
+from .file import  File
+from .chart import Chart
+from .associations import SharedFile, SharedChart
