+++ app/models/base.py
@@ -1,0 +1,10 @@
+from datetime import datetime, timezone
+
+from ..extensions import db
+
+
+class Base(db.Model):
+    __abstract__ = True
+    id = db.Column(db.Integer, primary_key=True)
+    created_at = db.Column(db.DateTime, default=datetime.now(timezone.utc))
+    updated_at = db.Column(db.DateTime, default=datetime.now(timezone.utc), onupdate=datetime.now(timezone.utc))
