+++ app/services/plots/__init__.py
@@ -1,0 +1,2 @@
+from .helpers import read_csv, save_to_string
+from .plotters import registry
