+++ app/services/plots/registry.py
