+++ requirements.txt
@@ -26,3 +26,4 @@ Werkzeug==3.1.3
 WTForms==3.2.1
 email-validator==2.1.1
 Flask-SQLAlchemy==3.1.1
+Flask-Migrate~=4.1.0
